Adds a 34 character limit to displayName input, displays an error if the user enters a name that is too long, added some styling classes to handle and break up long names when viewed in the settings page (linked accounts), improved the welcome message centering
