+++ src/shared/features/PostsView/PostsView.css
@@ -1,0 +1,3 @@
+.welcomeName {
+  text-align:center;
+}
