+++ src/shared/features/SettingsView/components/LinkedAccountManagement/LinkedAccountManagement.css
@@ -1,0 +1,7 @@
+.breakLongName {
+  word-wrap:break-word;
+}
+
+.noRight {
+  right:0px !important;
+}
